WIP didNavigate hook
The intention here is to provide a didNavigate()
hook on the route that is invoked whenever you
navigate to a route. The goal is that it would
receive parameters specific to a single action/
transition attempt.

I ended up not liking the API because it involves
all the same confusions in react surrounding
when willReceiveProps (and the new static
getDerivedStateFromProps) and having to diff
yourself, etc.

For context

https://twitter.com/amatchneer/status/1172921502674690048